requirements.txt
fastapi
uvicorn
faker
pydantic
sqlparse
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
Faker==37.6.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
